Wishlist With Saved Searches Added.
This took me a while to work out and could be improved but I think its kind of cool. I added a save button to each search. All the dog info of the saved search card is saved in localstorage. You can click the wishlist link and it will clear the display area and display your saved searches.